requirements.txt
torch
torchaudio
librosa
numpy
scipy
pandas
scikit-learn
matplotlib
seaborn
pyyaml
tqdm
